isr_project_be/requirements.txt
absl-py==2.0.0
amqp==5.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ecs-logging==2.0.2
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-curator==8.0.8
elasticsearch8==8.8.2
es_client==8.8.2.post1
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
kombu==5.3.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
prompt-toolkit==3.0.39
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.36.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0
vine==5.1.0
voluptuous==0.14.1
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.14.1
